Refactor foldContentLine into a separate class.
